Changes to reduce update latency
Modify handleGPS to operate on a copy of the buffer. Before calling it,
make a local copy of the buffer and release the volatile one back to
the ISR.
Redo interrupt priorities - the display update is now low, serial is
medium and the PPS capture is high.
Swap the loop structure of the display raster - do the digits inside so
that updating them with new contents is faster. Also, allow the raster
counters to be externally reset to reduce critical display update
latency when it counts.
Add support for using pin C7 as a marker for the end of the PPS ISR.
This allows measuring the PPS triggered display update latency.